adding a response for the entry in the databse which is not present an delete it if present

+++ src/main/java/org/example/service/DefaultCustomerService.java
@@ -8,6 +8,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -36,11 +37,18 @@ public class DefaultCustomerService implements CustomerService {
        return new ResponseEntity<Customer>(optionalCustomer.get(), HttpStatus.OK);
    }
 
-       return  new ResponseEntity<>(HttpStatus.NOT_FOUND);
+     throw new ResponseStatusException(HttpStatus.NOT_FOUND,"customer doest not exists");
     }
-    public void deleteCustomer(Long id)
-    {
-        repository.deleteById(id);
+    public void deleteCustomer(Long id){
+        //before deleting, verify that customer is in DB.
+        Optional<Customer> optionalCustomer=repository.findById(id);
+        if(optionalCustomer.isPresent()) {
+            repository.deleteById(id);
+//            return new ResponseEntity<>(optionalCustomer.get(),HttpStatus.OK);
+            throw new ResponseStatusException(HttpStatus.OK,"customer succesfully deleted");
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND,"customer id can't be deleted");
+
     }
